Tests: reworked libgd version detection.
The "libgd-config" binary is deprecated in recent versions and may not exist
or have unexpected output.  More, it may not present within older versions,
as well, if installed separately, which previously broke test assumptions.
The fix is change the fallback to skip tests.  In addition, recent Perl GD
module (2.57) started to export libgd version, which is now also consulted.

--HG--
branch : nginx-tests